server/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1
django-ckeditor==6.4.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-js-asset==2.0.0
djangorestframework==3.13.1
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
pillow==10.4.0
promise==2.3
psycopg2-binary==2.9.9
python-decouple==3.6
pytz==2022.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
whitenoise==6.2.0
